RandomMeas

Documentation for RandomMeas.jl: The randomized measurements toolbox in julia

Acquisition routines

Postprocessing routines

Postprocessing routines for classical shadows

Routines for simulating quantum circuits

Additional useful routines for ITensor

Acquisition routines

RandomMeas.get_Born — Method

 get_Born(ρ::MPO)

Construct Born Probability vector P(s) from an MPO representation ρ

RandomMeas.get_Born — Method

get_Born(ψ::MPS)

Construct Born Probability vector P(s)=|ψ(s)|^2 from an MPS representation ψ

RandomMeas.get_Born_MPS — Method

get_Born_MPS(ρ::MPO)

Construct Born Probability vector P(s)=<s|ρ|s> as an MPS from an MPO representation ρ

RandomMeas.get_Born_MPS — Method

get_Born_MPS(ψ::MPS)

Construct Born Probability vector P(s)=|ψ(s)|^2 as an MPS from an MPS representation ψ

RandomMeas.get_RandomMeas — Method

get_RandomMeas(ρ::Union{MPO,MPS}, u::Vector{ITensor})

Sample randomized measurements from a MPS/MPO representation ρ

RandomMeas.get_RandomMeas_MPO — Method

get_RandomMeas_MPO

Sample randomized measurements from an MPO representation ρ. The sampling is based from the MPO directly, i.e is memory-efficient

RandomMeas.get_RandomMeas_MPS — Method

get_RandomMeas_MPS(ψ::MPS, u::Vector{ITensor},NM::Int64)

Sample randomized measurements from an MPS representation ψ. The sampling is based from the MPS directly, i.e is memory-efficient

RandomMeas.get_Samples_Flat — Method

get_Samples_Flat(state::Union{MPO,MPS},NM::Int64)

Sample randomized measurements from a MPS/MPO representation ρ

RandomMeas.get_rotation — Method

get_rotation(ξ::Index{Int64}, cat::Int)

Generate a single qubit unitary with indices (ξ',ξ) Categories specified by cat: #1: Haar #2: Pauli #6: Identity matrix

RandomMeas.get_rotations — Function

get_rotations!

Generate a list of N single qubit unitaries to be applied on each qubit

Postprocessing routines

RandomMeas.get_Born — Method

get_Born(data::Array{Int8},ξ::Vector{Index{Int64}})

Construct histogram from randomized measurements as an ITensor representing the estimated Born probability

RandomMeas.get_overlap — Method

get_overlap(prob1::ITensor,prob2::ITensor,ξ::Vector{Index{Int64}},N::Int64)

Extract an overlap estimation $\mathrm{tr}(\rho_1\rho_2)$ from two estimated Born probabilities [See Elben et al 2019: Cross Platform verification of intermediate scale quantum devices]

RandomMeas.get_purity_hamming — Method

get_purity_hamming(data::Array{Int8},ξ::Vector{Index{Int64}})

Extract the purity from the Hamming distance formula purity = $\sum_s (-2)^{-D[s,s']}P(s)P(s')$ [Brydges et al, Science 2019]

RandomMeas.get_purity_shadows — Method

get_purity_shadows(data::Array{Int8},ξ::Vector{Index{Int64}})

Extract the purity from classical shadows

Postprocessing routines for classical shadows

RandomMeas.get_batch_shadows — Method

get_batch_shadows(data::Array{Int8}, ξ::Vector{Index{Int64}}, u::Vector{Vector{ITensor}}, n::Int64)

Constructs n batch shadows from measured data

RandomMeas.get_expect_shadow — Method

get_expect_shadow(O, shadow, ξ::Vector{Index{Int64}})

Contract shadow with operator O to estimate the expectation value $\mathrm{tr}(O\rho)$

RandomMeas.get_moments — Method

get_moments(shadow::Vector{ITensor}, ξ::Vector{Index{Int64}})

Obtain trace moments from a vector of (batch) shadows using U-statistics

RandomMeas.get_shadow — Method

get_shadow(P::ITensor, ξ::Vector{Index{Int64}}, u::Vector{ITensor};G::Union{Vector{Float64},Nothing}=nothing)

Form shadow from Born probability represented as an ITensor

RandomMeas.get_shadow_factorized — Method

get_shadow_factorized(data::Array{Int8}, s::Vector{Index{Int64}}, u::Vector{};G_vec::Union{Nothing,Vector{Float64}}=nothing)

build shadow as a tensor-product (memory-efficient)

RandomMeas.multiply — Method

multiply(shadow::ITensor, shadow2::ITensor)

RandomMeas.power — Method

power(shadow::ITensor, n::Int64)

RandomMeas.square — Method

square(shadow::ITensor)

RandomMeas.trace — Method

trace(shadow::ITensor, ξ::Vector{Index{Int64}})

Routines for simulating quantum circuits

RandomMeas.Apply_depo_channel — Method

Apply_depo_channel(ρ::MPO,p::Vector{Float64})

Apply depolarization channel on all qubits with probabilities specficied by the vector p

RandomMeas.RandomCircuit — Method

RandomCircuit(ξ::Vector{Index{Int64}},depth::Int64)

Create a random circuit of given depth. Returns the list of gates as a vector of ITensors

Additional useful routines for ITensor

RandomMeas.flatten — Method

flatten(O::Union{MPS,MPO,Vector{ITensor}})

convert a MPS/MPO to ITensor by a brutal multiplication A[1] x A[2] x ...

RandomMeas.get_entropy — Method

get_entropy(spec::ITensor)

compute von Neumann entropy from entanglement spectrum

RandomMeas.get_moment — Method

get_moment(spec::ITensor,n::Int)

compute the moments ``\mathrm{tr}(\rho)'' from the entanglement spectrum

RandomMeas.get_purity — Method

get_purity(ρ::MPO,part::Vector{Int64},ξ::Vector{Index{Int64}})

Compute the purity over the sites of part

RandomMeas.get_purity — Method

get_purity(ρ::MPO)

compute the purity of an MPO

RandomMeas.get_purity — Method

get_purity(ψ::MPS,i::Int64,j::Int64)

compute the purity over sites i:j

RandomMeas.get_purity — Method

get_purity(ψ::MPS,NA::Int64)

compute the purity of an MPS over the first NA sites

RandomMeas.get_spectrum — Method

get_spectrum(ψ::MPS,NA::Int64)
Compute entanglement spectrum of the first NA sites density matrix

RandomMeas.reduce_dm — Method

reduce_dm(ρ::MPO,i::Int64,j::Int64)

compute the reduced density matrix for sites i:j

RandomMeas.reduce_dm — Method

reduce_dm(ρ::MPO,part::Vector{Int64})

compute the reduce density matrix over sites mentionned in part

RandomMeas.reduce_dm — Method

reduce_dm(ψ::MPS,i::Int64,j::Int64)

compute the reduced density matrix for sites i:j

RandomMeas.trace — Method

trace(ρ::MPO,s::Vector{Index{Int64}})